fixed up the exit codes, and added addition functionality

+++ adduser.sh
@@ -20,34 +20,35 @@ while [ "$(($#))" != 0 ] && [ "$(($#))" != 5 ]; do
 	case $1 in
 	-a ) 
 		sudo_check ; 
-			if [ "$3" = --no-home ]; then
+			if [ "$3" = --no-home ] || [ "$3" = -nh ]; then
 				sudo useradd -M "$2"
 				echo "User $2 was created with no home directory"
-				exit 1
+				exit 0
 			fi
 		sudo useradd -m "$2" ;
 		echo "$2 was created with a home directory" ;
-	        exit 1 	
+	        exit 0 	
 		;;
 	-h ) 
 		echo "Usage: adduser <option> <username> <argument> <modifier>" ;
 		echo "---------------------------" ;
 		echo "OPTIONS:" ;
-		echo "-a <username>: adds a user" ;
+		echo "-a <username>: adds a user, by default a home directory is made" ;
 		echo "-aG <username> <group>: Creates a user to a group" ;
-		echo "-d <username>: deletes a user" ;
+		echo "-d <username>: deletes a user, by default the home directory is not deleted" ;
 		echo "-p <username>: changes a user password" ;
 		echo "-id or --change-id <user id> <username>: changes an existing user's UID" ;
 		echo "---------------------------" ;
 		echo "Modifiers:" ;
-		echo "-a: --no-home, creates a user with no home directory" ;
-		exit 1 
+		echo "-a: -nh --no-home, creates a user with no home directory" ;
+		echo "-d: -h --home, deletes a user with their home directory" ;
+		exit 0 
 		;;
 	-aG )	
 		sudo_check ;
 		sudo useradd -m "$2" -G "$3" ;
 		echo "$2 is added to group $3" ;
-		exit 1 
+		exit 0
 		;;
 	-id )
 		sudo_check ;
@@ -55,15 +56,20 @@ while [ "$(($#))" != 0 ] && [ "$(($#))" != 5 ]; do
 		;;
 	-d )  
 		sudo_check ;
-		sudo userdel -r "$2" ;
+			if [ "$3" = --home ] || [ "$3" = -h ]; then
+				sudo userdel -r "$2"
+				echo "$2 has been deleted with their home directory"
+				exit 0
+			fi
+		sudo userdel "$2" ;
 		echo "$2 has been deleted" ; 
-		exit 1
+		exit 0
 		;;
 	-p )  
 		sudo_check ;
 		sudo passwd "$2" ;
 		echo "$2's password has changed" ;
-		exit 1 
+		exit 0 
 		;;
 	esac
 done
